Decide on transaction date during GoCardless transaction normalization
Previously the Actual backend would default to using `valueDate` and
falling back to `bookingDate`.  According to the GoCardless
documentation[0] we should prefer `bookingDate` if it's set, but the
different bank integrations vary greatly in how they use the date fields
and so for consistency we should instead let the Actual bank
integrations decide which is the correct date for any given transaction.

For any transaction without an appropriate date to return the
`normalizeTransaction` function should return `null` so that it's
filtered out until the bank provides the full transaction details.

The existing bank integrations for ING_PL_INGBPLPW and
MBANK_RETAIL_BREXPLPW were changed to retain the existing precedence
order, though I'm not sure which is the correct one for these banks.

  [0]: https://nordigen.zendesk.com/hc/en-gb/articles/7899367372829-valueDate-and-bookingDate-for-transactions

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>