Add gocardless support for Sparkasse Karlsruhe (Germany)
Adds a gocardless formatter, to properly format transaction information.
Previously, additional remittance information was not supplied at all, and some
creditor names were not set. This made recognizing transactions hard for
users.

Now, we pick out the best remittanceInformation that we find, and use a
helpful order of possible data from the transaction as the creditorName.

Fixes #331