link to current relevant docs on init
Can't hurt (too much) to link to relevant documentation (until the link falls out of date...)

This was an AI-suggested change, for what it's worth.
Except that the reasonable-looking Docker docs link provided by the AI was incorrect so... I guess I'll take the credit myself for having fact-checked it.